api/requirements.txt
asgiref==3.7.2
bibtexparser==1.4.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2.1
django-cors-headers==4.0.0
django-filebrowser==4.0.2
django-grappelli==3.0.6
django-storages==1.13.2
django-tinymce==3.6.1
djangorestframework==3.14.0
feedparser==6.0.10
filebrowser==1.1.4
Flask==2.3.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
networkx==3.1
numpy==1.24.3
Pillow==9.5.0
protobuf==4.23.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
pyzotero==1.5.9
requests==2.31.0
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
django-simple-captcha
pysolr